Action model is done
Added a model where nodes can be subreddit, submission, top comments or sub/thread comments. every edge hold a weight value which is equal to the number of other activities this activity has generated, which means the number of nodes under this node in a tree. example an edge from subreddit node to submission node has the weight which is equal to the number of all comments on this submission.